# feat: create custom query to generate the fees per day