Create a custom SQL query to generate the fees per day.
Old way execution time: `19s 451.477879ms`
New query execution time: `1s 294.656516ms`

Almost 19 times faster.